add rebuild options

+++ build.sh
@@ -1,5 +1,8 @@
 # Create build directory
-rm -rf cmake
+## flag "D" provided -> rebuild
+if [ $# -eq 1 ]; then
+	if [ $1 == "D" ]; then echo "Rebuilding..." && rm -rf cmake; fi
+fi
 mkdir -p cmake/build
 
 pushd cmake/build
